Projects/1-Q&A Chatbot/requirements.txt
langchain-groq
langchain-google-genai
langchain
python-dotenv
langchain_community
langchain_core
streamlit